+++ astropy/coordinates/__init__.py
@@ -16,5 +16,6 @@ from .transformations import *
 from .builtin_systems import *
 from .name_resolve import *
 from .matching import *
+from .old_builtin_systems_names import *  # TODO: remove this in next version, along with module file
 
 __doc__ += builtin_systems._transform_graph_docs