+++ astropy/coordinates/tests/test_name_resolve.py
@@ -15,7 +15,7 @@ from ...tests.helper import pytest
 
 from ..name_resolve import get_icrs_coordinates, NameResolveError, \
                            SESAME_DATABASE, _parse_response
-from ..builtin_systems import ICRSCoordinates
+from ..builtin_systems import ICRS
 from ...tests.helper import remote_data
 from ... import units as u
 
@@ -119,9 +119,9 @@ def test_names():
         icrs = get_icrs_coordinates("ngc 3642")
     except NameResolveError:
         ra,dec = _parse_response(_cached_ngc3642["all"])
-        icrs = ICRSCoordinates(ra, dec, unit=(u.degree, u.degree))
+        icrs = ICRS(ra, dec, unit=(u.degree, u.degree))
 
-    icrs_true = ICRSCoordinates("11h 22m 18.014s", "59d 04m 27.27s")
+    icrs_true = ICRS("11h 22m 18.014s", "59d 04m 27.27s")
     np.testing.assert_almost_equal(icrs.ra.degree, icrs_true.ra.degree, 3)
     np.testing.assert_almost_equal(icrs.dec.degree, icrs_true.dec.degree, 3)
 
@@ -129,9 +129,9 @@ def test_names():
         icrs = get_icrs_coordinates("castor")
     except NameResolveError:
         ra,dec = _parse_response(_cached_castor["all"])
-        icrs = ICRSCoordinates(ra, dec, unit=(u.degree, u.degree))
+        icrs = ICRS(ra, dec, unit=(u.degree, u.degree))
 
-    icrs_true = ICRSCoordinates("07h 34m 35.87s", "+31d 53m 17.8s")
+    icrs_true = ICRS("07h 34m 35.87s", "+31d 53m 17.8s")
     np.testing.assert_almost_equal(icrs.ra.degree, icrs_true.ra.degree, 3)
     np.testing.assert_almost_equal(icrs.dec.degree, icrs_true.dec.degree, 3)
 
@@ -146,10 +146,10 @@ def test_database_specify():
     for db in ["simbad", "vizier", "all"]:
         SESAME_DATABASE.set(db)
         try:
-            icrs = ICRSCoordinates.from_name(name)
+            icrs = ICRS.from_name(name)
         except NameResolveError:
             ra,dec = _cached_ngc3642[db]
-            icrs = ICRSCoordinates(ra, dec, unit=(u.degree, u.degree))
+            icrs = ICRS(ra, dec, unit=(u.degree, u.degree))
 
         time.sleep(1)
 
@@ -158,9 +158,9 @@ def test_database_specify():
     for db in ["simbad",  "all"]:
         SESAME_DATABASE.set(db)
         try:
-            icrs = ICRSCoordinates.from_name(name)
+            icrs = ICRS.from_name(name)
         except NameResolveError:
             ra,dec = _cached_castor[db]
-            icrs = ICRSCoordinates(ra, dec, unit=(u.degree, u.degree))
+            icrs = ICRS(ra, dec, unit=(u.degree, u.degree))
 
         time.sleep(1)